refactoring into an MVC pattern, also removing a stray file
