+++ package.json
@@ -4,8 +4,7 @@
   "main": "index.js",
   "scripts": {
     "dev": "vite",
-    "server": "node server.js",
-    "start": "concurrently \"npm run server\" \"npm run dev\"",
+    "server": "node src/server/server.js",
     "build": "vite build",
     "lint": "eslint .",
     "test": "jest",
